# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Lingua-Bloom Source Ingestion API
  version: 0.3.0
paths:
...
  /api/imports/{runId}/resume:
    post:
      summary: Manually resume an owner-controlled run after a retriable failure
      security:
        - teacherSession: []
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResumeSubmission' }
      responses:
        '202': { description: Resume accepted; the same run continues from its last checkpoint }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Run not found }
        '409': { description: Run is not retriable or the idempotency key conflicts }
  /api/imports/{runId}/dispatch:
    post:
      summary: Idempotently redispatch an owner-controlled stale accepted or processing run
      security:
        - teacherSession: []
      parameters:
        - $ref: "#/components/parameters/RunId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RedispatchSubmission" }
      responses:
        "202": { description: Redispatch accepted for the same run }
        "400": { $ref: "#/components/responses/InvalidInput" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { description: Run not found }
        "409": { description: Run is not stale, already has a draft, or the idempotency key conflicts }
        "503": { description: Claim persisted but event delivery failed; retry with the same key }
...
  /api/imports/{runId}/publish:
    post:
      summary: Publish an immutable lesson version after all gates pass
      security:
        - teacherSession: []
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublishSubmission' }
      responses:
        '201':
          description: Lesson version created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublishedVersion' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: The canonical publication-readiness gate returned one or more reasons
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicationBlocked' }
...
components:
  securitySchemes:
    teacherSession:
      type: apiKey
      in: cookie
      name: sb-auth-token
      description: Supabase SSR project-scoped auth cookie; the deployed cookie may be chunked
  parameters:
    RunId:
      in: path
      name: runId
      required: true
      schema: { type: string, minLength: 1 }
...
  responses:
    InvalidInput:
      description: Invalid source, unsupported MIME type, or both/neither source fields supplied
    Unauthenticated:
      description: Valid teacher session required
    Forbidden:
      description: Authenticated teacher does not own the requested resource
...
  schemas:
...
    RedispatchSubmission:
      type: object
      additionalProperties: false
      required: [idempotencyKey]
      properties:
        idempotencyKey: { type: string, minLength: 16, maxLength: 128 }
    ResumeSubmission:
      type: object
      additionalProperties: false
      required: [idempotencyKey]
      properties:
        idempotencyKey: { type: string, minLength: 16, maxLength: 128 }
...
    PublishSubmission:
      type: object
      additionalProperties: false
      properties:
        confirmPermanentPublicAccess:
          const: true
          description: Required by the server only when the lesson has never been published before
    PublicationBlocked:
      type: object
      additionalProperties: false
      required: [code, message, reasons]
      properties:
        code: { const: PUBLISH_BLOCKED }
        message: { type: string, minLength: 1 }
        reasons:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, minLength: 1 }
    PublishedVersion:
      type: object
      additionalProperties: false
      required: [lessonId, publicLessonId, version]
      properties:
        lessonId: { type: string }
        publicLessonId:
          type: string
          minLength: 22
          pattern: '^[A-Za-z0-9_-]+$'
        version: { type: integer, minimum: 1 }